+++ src/index.ts
@@ -1,9 +1,3 @@
-export * from './d3util';
-export * from './databrowser';
-export * from './layout_d3util';
-export * from './link_renderer';
-export * from './LinkRepresentation';
+export * from './d3utils';
+export * from './mainUtils';
 export * from './link';
-export * from './Parser';
-export * from './selectioninfo';
-export * from './ToolTip';
