+++ src/link/LinkRepresentation.ts
