+++ src/mainUtils/selectioninfo.ts
